requirements.txt
python-decouple==3.7
schedule==1.1.0
Telethon==1.26.1
